Change chrome.downloads.DownloadItem.{fileSize,bytesReceived,totalBytes} from longs to doubles for files > 2GB.
Review URL: https://codereview.chromium.org/132353002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@244404 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 6396b29769ef2593823152d4ed0e0fe5b556958f

+++ chrome/browser/extensions/api/downloads/downloads_api.cc
@@ -258,8 +258,8 @@ scoped_ptr<base::DictionaryValue> DownloadItemToJSON(
   json->SetBoolean(kPausedKey, download_item->IsPaused());
   json->SetString(kMimeKey, download_item->GetMimeType());
   json->SetString(kStartTimeKey, TimeToISO8601(download_item->GetStartTime()));
-  json->SetInteger(kBytesReceivedKey, download_item->GetReceivedBytes());
-  json->SetInteger(kTotalBytesKey, download_item->GetTotalBytes());
+  json->SetDouble(kBytesReceivedKey, download_item->GetReceivedBytes());
+  json->SetDouble(kTotalBytesKey, download_item->GetTotalBytes());
   json->SetBoolean(kIncognitoKey, profile->IsOffTheRecord());
   if (download_item->GetState() == DownloadItem::INTERRUPTED) {
     json->SetString(kErrorKey, content::InterruptReasonDebugString(
@@ -290,7 +290,7 @@ scoped_ptr<base::DictionaryValue> DownloadItemToJSON(
       json->SetString(kByExtensionNameKey, extension->name());
   }
   // TODO(benjhayden): Implement fileSize.
-  json->SetInteger(kFileSizeKey, download_item->GetTotalBytes());
+  json->SetDouble(kFileSizeKey, download_item->GetTotalBytes());
   return scoped_ptr<base::DictionaryValue>(json);
 }
